pyproject.toml
[project]
name = "mcplatex-gui"
version = "0.1.0"
description = "mcplatex with gui"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyside6-fluent-widgets[full]",
    "python-dotenv"
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true